requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==21.7
SQLAlchemy==2.0.23
alembic==1.13.1
httpx==0.27.0
python-dotenv==1.0.0
pydantic==1.10.13
redis==5.0.1
aioredis==2.0.1
